+++ tests/framework/console/controllers/FixtureControllerTest.php
@@ -57,6 +57,18 @@ class FixtureControllerTest extends TestCase
         $this->assertCount(1, FixtureStorage::$firstFixtureData, 'first fixture data should be loaded');
     }
 
+    public function testLoadGlobalFixtureWithFixture()
+    {
+        $this->_fixtureController->globalFixtures = [
+            '\yiiunit\data\console\controllers\fixtures\GlobalFixture',
+        ];
+
+        $this->_fixtureController->actionLoad(['First']);
+
+        $this->assertCount(1, FixtureStorage::$globalFixturesData, 'global fixture data should be loaded');
+        $this->assertCount(1, FixtureStorage::$firstFixtureData, 'first fixture data should be loaded');
+    }
+
     public function testUnloadGlobalFixture()
     {
         $this->_fixtureController->globalFixtures = [